refactor: seperate repository interface
