+++ src/user/user.module.ts
@@ -1,11 +1,18 @@
 import { Module } from '@nestjs/common';
 import { UserService } from './user.service';
 import { UserRepository } from './user.repository';
+import { UserPrismaRepository } from './user.prisma.repository';
 import { UserController } from './user.controller';
 
 @Module({
-  providers: [UserService, UserRepository],
-  exports: [UserRepository],
   controllers: [UserController],
+  providers: [
+    UserService,
+    {
+      provide: UserRepository,
+      useClass: UserPrismaRepository,
+    },
+  ],
+  exports: [UserRepository],
 })
 export class UserModule {}
